Update from original

+++ modules/activiti-engine/src/test/java/org/activiti/engine/test/bpmn/event/compensate/helper/UndoService.java
@@ -24,12 +24,12 @@ import org.activiti.engine.delegate.JavaDelegate;
 public class UndoService implements JavaDelegate {
   
   private Expression counterName;
-  
+
   public void execute(DelegateExecution execution) throws Exception {
     String variableName = (String) counterName.getValue(execution);
     Object variable = execution.getVariable(variableName);
     if(variable == null) {
-      execution.setVariable(variableName, (Integer) 1);
+      execution.setVariable(variableName, 1);
     }
     else  {
       execution.setVariable(variableName, ((Integer)variable)+1);
